+++ lib/console/commands/new.dart
@@ -4,19 +4,19 @@ import 'dart:io';
 import 'package:archive/archive_io.dart';
 import 'package:args/command_runner.dart';
 import 'package:crypto/crypto.dart';
-import 'package:daravel_core/console/logger.dart';
+import 'package:daravel_core/console/console_logger.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:path/path.dart' as path;
 import 'package:http/http.dart' as http;
 
 import 'generate.dart';
 
-class CreateCommand extends Command {
+class NewCommand extends Command {
   @override
   String get description => "Create a new Daravel project";
 
   @override
-  String get name => "create";
+  String get name => "new";
 
   static final Directory _cacheDir =
       Directory(path.join(_homeDirectory, ".daravel"));
@@ -27,9 +27,9 @@ class CreateCommand extends Command {
   final File _projectTemplateFileInfo =
       File(path.join(_cacheDir.path, "project_template.zip.json"));
 
-  late final Logger logger = Logger();
+  late final ConsoleLogger logger = ConsoleLogger();
 
-  CreateCommand() {
+  NewCommand() {
     argParser.addOption(
       "project-name",
       abbr: "p",