+++ lib/database/concerns/query_result.dart
@@ -1,5 +1,0 @@
-abstract class QueryResult {
-  List<List<Object?>> get rows;
-  List<Map<String, Object?>>? get mappedRows;
-  Object? get resultObject;
-}