+++ lib/database/db_connection.dart
@@ -1,6 +1,6 @@
 import 'package:daravel_core/daravel_core.dart';
 import 'package:daravel_core/database/concerns/db_driver.dart';
-import 'package:daravel_core/database/concerns/query_result.dart';
+import 'package:daravel_core/database/concerns/record_set.dart';
 import 'package:daravel_core/database/drivers/sqlite/sqlite.dart';
 
 class DBConnection {
@@ -16,7 +16,7 @@ class DBConnection {
     }
   }
 
-  QueryResult? select(String query, [List<dynamic> bindings = const []]) {
+  RecordSet? select(String query, [List<dynamic> bindings = const []]) {
     return driver.select(query, bindings);
   }
 
@@ -28,12 +28,12 @@ class DBConnection {
     return driver.insert(query, bindings);
   }
 
-  bool delete(String query, [List<dynamic> bindings = const []]) {
+  Future<int> delete(String query, [List<dynamic> bindings = const []]) {
     return driver.delete(query, bindings);
   }
 
-  bool update(String query, [List<dynamic> bindings = const []]) {
-    return driver.update(query, bindings);
+  Future<int> update(String query, [List<dynamic> bindings = const []]) async {
+    return await driver.update(query, bindings);
   }
 
   bool unprepared(String query) {