+++ test/dartisan_test.dart
@@ -4,9 +4,9 @@ import 'dart:io';
 import 'package:test/test.dart';
 import 'package:path/path.dart' as path;
 
-import '../bin/src/commands/create.dart';
-import '../bin/src/commands/generate.dart';
-import '../bin/src/commands/make_config.dart';
+import 'package:daravel_core/console/commands/new.dart';
+import 'package:daravel_core/console/commands/generate.dart';
+import 'package:daravel_core/console/commands/make_config.dart';
 
 void main() {
   tearDown(() {
@@ -40,7 +40,7 @@ void main() {
       ),
     );
 
-    expect(logs, ['\x1B[31m', '[ERROR] Config directory not found.']);
+    expect(logs, ['\x1B[31m[ERROR] \x1B[37mConfig directory not found.']);
     logs.clear();
 
     if (!playgroundConfigDirectory.existsSync()) {
@@ -132,7 +132,7 @@ void main() {
       playgroundDirectory.createSync();
     }
 
-    await CreateCommand().run(playgroundDirectory.path, 'test_project');
+    await NewCommand().run(playgroundDirectory.path, 'test_project');
 
     final projectDirectory =
         Directory(path.join(playgroundDirectory.path, 'test_project'));
@@ -143,7 +143,7 @@ void main() {
 
     await runZonedGuarded(
       () async {
-        await CreateCommand().run(playgroundDirectory.path, 'test_project');
+        await NewCommand().run(playgroundDirectory.path, 'test_project');
       },
       (e, s) {},
       zoneSpecification: ZoneSpecification(
@@ -153,13 +153,13 @@ void main() {
       ),
     );
 
-    expect(
-        logs, ['\x1B[33m', '[WARNING] Directory test_project already exists']);
+    expect(logs,
+        ['\x1B[33m[WARNING] \x1B[37mDirectory test_project already exists']);
     logs.clear();
 
     await runZonedGuarded(
       () async {
-        await CreateCommand().run(
+        await NewCommand().run(
           playgroundDirectory.path,
         );
       },
@@ -171,7 +171,7 @@ void main() {
       ),
     );
 
-    expect(logs, ['\x1B[33m', '[WARNING] Please provide a project name']);
+    expect(logs, ['\x1B[33m[WARNING] \x1B[37mPlease provide a project name']);
   });
 
   test('Generate Config File', () async {
@@ -185,7 +185,7 @@ void main() {
       playgroundDirectory.createSync();
     }
 
-    await CreateCommand()
+    await NewCommand()
         .run(playgroundDirectory.path, 'make_config_test_project');
 
     final projectDirectory = Directory(
@@ -202,7 +202,7 @@ void main() {
             .existsSync(),
         true);
 
-    await CreateCommand()
+    await NewCommand()
         .run(playgroundDirectory.path, 'error_make_config_test_project');
 
     await runZonedGuarded(
@@ -219,7 +219,7 @@ void main() {
       ),
     );
 
-    expect(logs, ['\x1B[31m', '[ERROR] Please provide config file name.']);
+    expect(logs, ['\x1B[31m[ERROR] \x1B[37mPlease provide config file name.']);
     logs.clear();
 
     await runZonedGuarded(
@@ -235,6 +235,6 @@ void main() {
       ),
     );
 
-    expect(logs, ['\x1B[31m', '[ERROR] Config directory not found.']);
+    expect(logs, ['\x1B[31m[ERROR] \x1B[37mConfig directory not found.']);
   });
 }