+++ test/database/sqlite/driver_test.dart
@@ -1,16 +1,18 @@
 import 'dart:io';
 
-import 'package:daravel_core/daravel_core.dart';
-import 'package:daravel_core/database/concerns/query_result.dart';
-
-import 'package:daravel_core/database/schema.dart';
+import 'package:path/path.dart' as path;
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
-import 'package:path/path.dart' as path;
+
+import 'package:daravel_core/daravel_core.dart';
+import 'package:daravel_core/database/concerns/record_set.dart';
+import 'package:daravel_core/exceptions/component_not_booted.dart';
+import 'package:daravel_core/exceptions/db_connection_not_found.dart';
 
 void main() {
   setUpAll(() {
-    Directory(path.join(Directory.current.path, 'test/database-playground'))
+    Directory(path.join(
+            Directory.current.path, 'test/database-sqlite-driver-playground'))
         .createSync(recursive: true);
 
     expect(() => DB.connection(), throwsA(isA<ComponentNotBootedException>()));
@@ -20,20 +22,20 @@ void main() {
       'database.connections': {
         'sqlite': DatabaseConnection(
           driver: 'sqlite',
-          database: 'test/database-playground/database.sqlite',
+          database: 'test/database-sqlite-driver-playground/database.sqlite',
           prefix: '',
           foreignKeyConstraints: true,
         ),
         'sqlite1': DatabaseConnection(
           driver: 'sqlite',
-          database: 'test/database-playground/database1.sqlite',
+          database: 'test/database-sqlite-driver-playground/database1.sqlite',
           prefix: '',
           foreignKeyConstraints: true,
           busyTimeout: 5000,
         ),
         'sqlite2': DatabaseConnection(
           driver: 'sqlite',
-          database: 'test/database-playground/database2.sqlite',
+          database: 'test/database-sqlite-driver-playground/database2.sqlite',
           prefix: '',
           foreignKeyConstraints: true,
         ),
@@ -47,7 +49,8 @@ void main() {
   });
 
   tearDownAll(() {
-    Directory(path.join(Directory.current.path, 'test/database-playground'))
+    Directory(path.join(
+            Directory.current.path, 'test/database-sqlite-driver-playground'))
         .deleteSync(recursive: true);
   });
 
@@ -319,16 +322,16 @@ void main() {
     // Select from connection.
     var result = DB.connection()!.select('SELECT * FROM $table');
 
-    expect(result, isA<QueryResult>());
+    expect(result, isA<RecordSet>());
 
-    expect(result!.rows.length, 0);
+    expect(result!.length, 0);
 
     // Direct Select
     result = DB.select('SELECT * FROM $table');
 
-    expect(result, isA<QueryResult>());
+    expect(result, isA<RecordSet>());
 
-    expect(result!.rows.length, 0);
+    expect(result!.length, 0);
   });
 
   test('Secondary Sqlite connection', () {
@@ -341,9 +344,9 @@ void main() {
     // Select from connection.
     final result = DB.connection('sqlite1')!.select('SELECT * FROM $table');
 
-    expect(result, isA<QueryResult>());
+    expect(result, isA<RecordSet>());
 
-    expect(result!.rows.length, 0);
+    expect(result!.length, 0);
   });
 
   test('Test non-existent connection & switching of default connection', () {
@@ -361,9 +364,9 @@ void main() {
 
     final result = DB.select('SELECT * FROM $table');
 
-    expect(result, isA<QueryResult>());
+    expect(result, isA<RecordSet>());
 
-    expect(result!.rows.length, 0);
+    expect(result!.length, 0);
   });
 
   test('Insert statement', () {
@@ -381,18 +384,18 @@ void main() {
 
     final selectResult = DB.select('SELECT * FROM $table');
 
-    expect(selectResult!.rows.length, 1);
+    expect(selectResult!.length, 1);
 
-    expect(selectResult.rows.first[0], 1);
-    expect(selectResult.rows.first[1], 'john@gmail.com');
-    expect(selectResult.rows.first[2], 'password');
+    expect(selectResult.first['id'], 1);
+    expect(selectResult.first['email'], 'john@gmail.com');
+    expect(selectResult.first['password'], 'password');
 
-    expect(selectResult.mappedRows!.first['id'], 1);
-    expect(selectResult.mappedRows!.first['email'], 'john@gmail.com');
-    expect(selectResult.mappedRows!.first['password'], 'password');
+    expect(selectResult.first['id'], 1);
+    expect(selectResult.first['email'], 'john@gmail.com');
+    expect(selectResult.first['password'], 'password');
   });
 
-  test('Delete statement', () {
+  test('Delete statement', () async {
     final table = 'users_17';
 
     Schema.create(table, (table) {
@@ -408,19 +411,19 @@ void main() {
 
     var selectResult = DB.select('SELECT * FROM $table');
 
-    expect(selectResult!.rows.length, 1);
+    expect(selectResult!.length, 1);
 
-    var result =
-        DB.delete('DELETE FROM $table WHERE email = ?', ['frank@gmail.com']);
+    var result = await DB
+        .delete('DELETE FROM $table WHERE email = ?', ['frank@gmail.com']);
 
-    expect(result, true);
+    expect(result, 1);
 
     selectResult = DB.select('SELECT * FROM $table');
 
-    expect(selectResult!.rows.length, 0);
+    expect(selectResult!.length, 0);
   });
 
-  test('Update statement', () {
+  test('Update statement', () async {
     final table = 'users_18';
 
     Schema.create(table, (table) {
@@ -436,26 +439,26 @@ void main() {
 
     var selectResult = DB.select('SELECT * FROM $table');
 
-    expect(selectResult!.rows.length, 1);
+    expect(selectResult!.length, 1);
 
-    var result = DB.update(
+    var result = await DB.update(
       'UPDATE $table SET email = ?, password = ? WHERE email = ?',
       ['john-edited@gmail.com', 'new-password', 'john@gmail.com'],
     );
 
-    expect(result, true);
+    expect(result, 1);
 
     selectResult = DB.select('SELECT * FROM $table');
 
-    expect(selectResult!.rows.length, 1);
+    expect(selectResult!.length, 1);
 
-    expect(selectResult.rows.first[0], 1);
-    expect(selectResult.rows.first[1], 'john-edited@gmail.com');
-    expect(selectResult.rows.first[2], 'new-password');
+    expect(selectResult.first['id'], 1);
+    expect(selectResult.first['email'], 'john-edited@gmail.com');
+    expect(selectResult.first['password'], 'new-password');
 
-    expect(selectResult.mappedRows!.first['id'], 1);
-    expect(selectResult.mappedRows!.first['email'], 'john-edited@gmail.com');
-    expect(selectResult.mappedRows!.first['password'], 'new-password');
+    expect(selectResult.first['id'], 1);
+    expect(selectResult.first['email'], 'john-edited@gmail.com');
+    expect(selectResult.first['password'], 'new-password');
   });
 
   test('Rename table', () {
@@ -471,7 +474,7 @@ void main() {
 
     final result = DB.select('SELECT * FROM new_users');
 
-    expect((result!.resultObject as ResultSet).length, 0);
+    expect(result!.length, 0);
   });
 
   test('Drop table', () {